remove unused, same window
