+++ src/helpers/constants.ts
@@ -17,7 +17,6 @@ export enum Screens {
     Beneficiary = 'Beneficiary',
     CommunityManager = 'CommunityManager',
     Communities = 'Communities',
-    CommunitiesList = 'CommunitiesList',
     Profile = 'Profile',
     Auth = 'Auth',
 }
