+++ src/views/communities/Stories.tsx
@@ -1,22 +1,18 @@
 import { Screens } from 'helpers/constants';
 import { navigationRef } from 'helpers/rootNavigation';
 import React, { Component } from 'react';
-import { ScrollView, Text, View, Image, Pressable } from 'react-native';
+import { ScrollView, Text, View } from 'react-native';
 import { Headline } from 'react-native-paper';
+import NewStoryCard from './NewStoryCard';
+import StoriesCard from './StoriesCard';
 
 export default class Stories extends Component<{}, {}> {
     render() {
         return (
-            <View
-                style={{
-                    // backgroundColor: 'red',
-                    marginHorizontal: 18,
-                    marginBottom: 18,
-                }}
-            >
+            <View>
                 <View
                     style={{
-                        // backgroundColor: 'red',
+                        marginHorizontal: 18,
                         flexDirection: 'row',
                         justifyContent: 'space-between',
                         alignItems: 'flex-end',
@@ -35,83 +31,12 @@ export default class Stories extends Component<{}, {}> {
                 <ScrollView
                     horizontal={true}
                     showsHorizontalScrollIndicator={false}
+                    contentContainerStyle={{ padding: 18 }}
                 >
-                    <Pressable
-                        style={{
-                            backgroundColor: 'blue',
-                            width: 98.16,
-                            height: 167,
-                            borderRadius: 8,
-                            marginRight: 11.84,
-                            // flexDirection: 'row',
-                            // alignItems: 'flex-end',
-                        }}
-                        onPress={(e) =>
-                            navigationRef.current?.navigate(
-                                Screens.StoriesCarousel
-                            )
-                        }
-                    >
-                        <Image
-                            source={{ uri: 'https://reactjs.org/logo-og.png' }}
-                            style={{
-                                flex: 1,
-                                // resizeMode: 'cover',
-                                justifyContent: 'center',
-                                // alignItems: 'flex-end',
-                                borderRadius: 8,
-                            }}
-                        />
-                        <View
-                            style={{
-                                flexDirection: 'row',
-                                alignItems: 'flex-end',
-                            }}
-                        >
-                            <Text
-                                style={{
-                                    backgroundColor: 'red',
-                                    position: 'absolute',
-                                    fontFamily: 'Gelion-Bold',
-                                    fontSize: 14,
-                                    lineHeight: 16,
-                                    color: '#FFFFFF',
-                                    paddingBottom: 9,
-                                    marginHorizontal: 12,
-                                }}
-                            >
-                                Toze
-                            </Text>
-                        </View>
-                        {/* </ImageBackground> */}
-                    </Pressable>
-                    <View
-                        style={{
-                            backgroundColor: 'blue',
-                            width: 98.16,
-                            height: 167,
-                            borderRadius: 8,
-                            marginRight: 11.84,
-                        }}
-                    ></View>
-                    <View
-                        style={{
-                            backgroundColor: 'blue',
-                            width: 98.16,
-                            height: 167,
-                            borderRadius: 8,
-                            marginRight: 11.84,
-                        }}
-                    ></View>
-                    <View
-                        style={{
-                            backgroundColor: 'blue',
-                            width: 98.16,
-                            height: 167,
-                            borderRadius: 8,
-                            marginRight: 11.84,
-                        }}
-                    ></View>
+                    <NewStoryCard />
+                    <StoriesCard />
+                    <StoriesCard />
+                    <StoriesCard />
                 </ScrollView>
             </View>
         );
